backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.54
alembic==1.12.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
google-generativeai==0.3.2
python-dotenv==1.1.0
firebase-admin==6.9.0
google-cloud-texttospeech==2.27.0
pydantic==1.10.12
pytest==8.4.0
httpx==0.28.1 